requirements.txt
Flask
gunicorn 
pandas
numpy
sklearn
datetime
ipykernel
pyYAML
-e .